# api/OpenAPI/comment.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Comments API",
    "version": "1.0.0",
    "description": "API for managing comments related to content."
  },
  "paths": {
...
    "/api/comments/delete/{id}": {
      "delete": {
        "summary": "Delete a specific comment",
        "operationId": "deleteComment",
        "tags": ["comments"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the comment to delete",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Comment deleted successfully"
          },
          "404": {
            "description": "Comment not found"
          }
        }
      }
    }
  },
...
}